Inline PML-CM

Conflicts:
	ompi/mca/mtl/mtl.h
