+++ ompi/mca/mtl/mtl.h
@@ -1,7 +1,10 @@
+/* -*- Mode: C; c-basic-offset:4 ; indent-tabs-mode:nil -*- */
 /*
  * Copyright (c) 2004-2006 The Regents of the University of California.
  *                         All rights reserved.
  * Copyright (c) 2012      Sandia National Laboratories.  All rights reserved.
+ * Copyright (c) 2015      Los Alamos National Security, LLC. All rights
+ *                         reserved.
  * $COPYRIGHT$
  * 
  * Additional copyrights may follow
@@ -58,7 +61,7 @@ typedef struct mca_mtl_request_t mca_mtl_request_t;
  *
  * Initialization routine for MTL component.  This function should
  * allocate resources for communication and try to do all local setup.
- * It should not attempt to contract it's peers, as that should be
+ * It should not attempt to contact it's peers, as that should be
  * done at add_procs time.  Contact information should be published
  * during this initialization function.  It will be made available
  * during add_procs().
@@ -428,17 +431,17 @@ typedef struct mca_mtl_base_module_t mca_mtl_base_module_t;
  */
 #if MCA_ompi_mtl_DIRECT_CALL
 
-#include MCA_ompi_mtl_DIRECT_CALL_HEADER
 
 #define OMPI_MTL_CALL_STAMP(a, b) ompi_mtl_ ## a ## _ ## b
 #define OMPI_MTL_CALL_EXPANDER(a, b) OMPI_MTL_CALL_STAMP(a,b)
 #define OMPI_MTL_CALL(a) OMPI_MTL_CALL_EXPANDER(MCA_ompi_mtl_DIRECT_CALL_COMPONENT, a)
 
+#include MCA_ompi_mtl_DIRECT_CALL_HEADER
+
 #else
 #define OMPI_MTL_CALL(a) ompi_mtl->mtl_ ## a
 #endif
 
-OMPI_DECLSPEC extern mca_mtl_base_module_t *ompi_mtl;
 
 END_C_DECLS
 #endif
