+++ ompi/mca/pml/pml.h
@@ -65,6 +65,7 @@
 #include "ompi_config.h"
 #include "opal/mca/mca.h"
 #include "mpi.h" /* needed for MPI_ANY_TAG */
+#include "ompi/mca/pml/pml_constants.h"
 
 BEGIN_C_DECLS
 
@@ -74,20 +75,6 @@ BEGIN_C_DECLS
 
 typedef uint64_t mca_pml_sequence_t;
 struct ompi_proc_t;
-
-typedef enum {
-    MCA_PML_BASE_SEND_SYNCHRONOUS,
-    MCA_PML_BASE_SEND_COMPLETE,
-    MCA_PML_BASE_SEND_BUFFERED,
-    MCA_PML_BASE_SEND_READY,
-    MCA_PML_BASE_SEND_STANDARD,
-    MCA_PML_BASE_SEND_SIZE
-} mca_pml_base_send_mode_t;
-
-
-#define OMPI_ANY_TAG    MPI_ANY_TAG
-#define OMPI_ANY_SOURCE MPI_ANY_SOURCE
-#define OMPI_PROC_NULL  MPI_PROC_NULL
 
 /**
  * MCA->PML Called by MCA framework to initialize the component.
